Content Helper Settings: Reduce code duplication
